+++ Physics/Source/Utility.cpp
@@ -85,8 +85,8 @@ std::string vtos(Vector3& v)
 }
 float AngleBetween(const Vector3 &a, const Vector3 &b)
 {
-	double num = a.Dot(b);
-	double den = a.Length() * b.Length();
+	float num = a.Dot(b);
+	float den = a.Length() * b.Length();
 	float RAngle = acos(num / den);
 	float DAngle = Math::RadianToDegree(RAngle);
 	return DAngle;
